+++ src/stdlib.rs
@@ -71,7 +71,7 @@ pub(crate) fn load_stdlib(context: &mut Rant)
 
   load_funcs!(
     // General functions
-    alt, call, cat, either, len, get_type as "type", seed, nop, range, resolve, fork, unfork, try_ as "try",
+    alt, call, cat, either, len, get_type as "type", seed, nop, range, irange, resolve, fork, unfork, try_ as "try",
 
     // Assertion functions
     assert as "assert", assert_eq as "assert-eq", assert_neq as "assert-neq",
@@ -116,7 +116,7 @@ pub(crate) fn load_stdlib(context: &mut Rant)
     proto, set_proto as "set-proto",
 
     // Collection functions
-    assoc, clear, collect, has, keys, index_of as "index-of", insert, last_index_of as "last-index-of", list, nlist, remove, sift, sifted, squish, squished, take, translate,
+    assoc, clear, collect, has, keys, index_of as "index-of", insert, last_index_of as "last-index-of", list, nlist, remove, rev, sift, sifted, squish, squished, take, translate,
 
     // List functions
     pick, filter, join, map, sort, sorted, shuffle, shuffled, sum,